// packages/plugin-moxie-token-transfer/src/types/types.ts
export interface CallbackTemplate {
    text: string;
    content?: {
        error?: string;
        details?: string;
        action?: string;
        inReplyTo?: string;
        confirmation_required?: boolean;
        confirmation_message?: string;
    };
}
export interface FunctionResponse<T> {
    callBackTemplate?: CallbackTemplate;
    data?: T;
}

export interface Context {
    traceId: string;
    moxieUserId?: string;
    runtime: IAgentRuntime;
    state: State;
    provider?: ethers.JsonRpcProvider;
    [key: string]: any;
}